% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_equal.R
\name{all_equal}
\alias{all_equal}
\title{Test a vector (numeric or otherwise) for equivalency across all elements}
\usage{
all_equal(x, tol=.Machine$double.eps ^ 0.5)
}
\arguments{
\item{x}{vector to test for equivalency}

\item{tol}{numeric, the tolerance to use for equivalency of numeric objects. Ignored if \code{x} is not numeric. Defaults to the square-root of the machine precision, typically about 1.5e-8.}
}
\value{
logical, either TRUE (all elements are equivalent) or FALSE (some elements are not equivalent)
}
\description{
Test a vector (numeric or otherwise) for equivalency across all elements. For numeric values, a tolerance is used to test equivalency. For non-numeric values, \code{unique} is used to determine equivalency.
}
