% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/big_colorblind_pal.R
\name{big_colorblind_pal}
\alias{big_colorblind_pal}
\title{Generate a large colorblind palette}
\usage{
big_colorblind_pal(
  n_colors,
  shuffle_colors = FALSE,
  drop_yellow = TRUE, drop_black = FALSE)
}
\arguments{
\item{n_colors}{integer, the number of colors in the palette. No default. Anything less than the number of colors in colorblind_pal (after dropping yellow and/or black) just returns the colorblind_pal colors.}

\item{shuffle_colors}{logical, whether to shuffle the palette order. Defaults to FALSE.}

\item{drop_yellow}{logical, whether to drop the difficult-to-see yellow color. Defaults to TRUE.}

\item{drop_black}{logical, whether to drop the black from the palette. Defaults to FALSE.}
}
\value{
a character vector with the colors
}
\description{
Generate a large colorblind palette, optionally with the colors shuffled to make adjacent colors more distinctive. This is a wrapper for colorblind_pal() and colorRampPalette(). It also drops the difficult-to-see yellow of colorblind_pal
}
