% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coalesce_with_checks.R
\name{coalesce_with_checks}
\alias{coalesce_with_checks}
\title{Collapse multiple redundant columns into a single column}
\usage{
coalesce_with_checks(x, columns)
}
\arguments{
\item{x}{data frame containing the redundant columns to be coalesced/collapsed}

\item{columns}{character vector, the names of the columns to be coalesced/collapsed. The values will be combined into the first column listed in \code{columns}, and the remaining columns will be dropped.}
}
\value{
A data frame with \code{columns} values collapsed into the first column listed.
}
\description{
Collapse multiple columns with redundant information into a single column. This function
is similar to dplyr::coalesce, but uses several checks and informative error messages.
Specifically, it checks for more than one unique non-NA value in each row, and returns an
error if there are any such rows (in similar cases, dplyr::coalesce uses the first
non-NA value). It then takes, for each row, the unique non-NA value found in one or
more of the input columns, and places that value into the first column.
}
