% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coerce_NA_text.R
\name{coerce_NA_text}
\alias{coerce_NA_text}
\title{Coerce textual NAs to true NA values}
\usage{
coerce_NA_text(x, NA_values=c("NA", "na", "N/A", "n/a", "", "."), verbose=FALSE)
}
\arguments{
\item{x}{vector, matrix, or data frame containing the data to be coerced}

\item{NA_values}{character vector, the values to be coerced to NAs.}

\item{verbose}{logical, whether to output information on data being coerced; if \code{x} is a data frame, the function will output the names of the columns with data being coerced.}
}
\value{
An object of the same type as the input data
}
\description{
This function is designed to convert textual NAs to true NA values. It can operate on vectors, data
frames, data frame like objects (e.g. tibbles), or matrices. It returns an object of the same class.
The values that will be coerced to NAs can be provided as a character vector.
}
