% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_not_NAs.R
\name{count_not_NAs}
\alias{count_not_NAs}
\title{Calculate the number of non-NA elements in an object}
\usage{
count_not_NAs(x)
}
\arguments{
\item{x}{object to be tested for non-NA values. Can be a data.frame, matrix, vector, list. For lists, it looks at the list elements themselves, not within those elements. For other object types, results may not be as expected.}
}
\value{
integer, the number of non-NA values in \code{x}
}
\description{
Calculate the number of non-NA elements in an object. Can be useful for determining the sparseness of matrices and data frames, or their component rows or columns.
}
