% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{cv}
\alias{cv}
\title{Calculate the coefficient of variation}
\usage{
cv(x, na.rm=FALSE)
}
\arguments{
\item{x}{numeric vector or matrix}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
a single numeric value
}
\description{
Calculate the coefficient of variation a vector of numbers, as \code{SD(x) / mean(x)}.
This function does NOT calculate percent CV - that requires multiplying the returned
values by 100.
}
