% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/factor_sorted.R
\name{factor_sorted}
\alias{factor_sorted}
\title{Convert a vector to a factor with levels sorted}
\usage{
factor_sorted(x, numeric = TRUE, ...)
}
\arguments{
\item{x}{vector to convert}

\item{numeric}{logical, passed to \code{stringr::str_sort}}

\item{...}{optional additional arguments passed to \code{stringr::str_sort}}
}
\value{
vector of the same length as x
}
\description{
This function converts a vector of any class to a factor with levels sorted according to specifications. It is a wrapper for \code{factor()} and \code{stringr::str_sort()}. By default, it sorts mixed character/numeric using the "numeric = TRUE" option for \code{str_sort}.
}
