% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_factors.R
\name{fix_factors}
\alias{fix_factors}
\title{Trim excess factor levels from columns of a data frame}
\usage{
\code{fix_factors(data)}
}
\arguments{
\item{data}{a data frame, with 0 or more columns of class \code{factor}}
}
\value{
a data frame of the same dimensions as \code{data}
}
\description{
This function goes through all columns of a data frame, and removes excess levels from columns of
class \code{factor}, while leaving other columns unchanged. It is intended to be used after
filtering a data frame, so that excess levels do not cause errors or confusion in downstream
processes.
}
