% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geomean.R
\name{geomean}
\alias{geomean}
\title{Calculate the geometric mean of a numeric vector}
\usage{
\code{geomean(x, ...)}
}
\arguments{
\item{x}{numeric vector. All elements must be positive, or an error is thrown. Geometric mean is defined only where all elements are positive.}

\item{...}{optional arguments, passed to \code{mean}. \code{na.rm} is particularly useful.}
}
\value{
numeric, the value of the geometric mean
}
\description{
Calculate the geometric mean of a numeric vector, passing arguments on to \code{mean}.
}

