% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nin.R
\name{\%in\%}
\alias{\%in\%}
\title{Determine non-matches of one vector to another}
\usage{
\code{x \%nin\% y}
}
\arguments{
\item{x}{vector to be matched to elements in \code{y}.}

\item{y}{vector to be searched for elements of \code{x}.}
}
\value{
a logical vector with one element for each element in x
}
\description{
Determine the members of one vector that are NOT found in another vector.  This is equivalent to
\code{!(x \%in\% y)}, but is a useful shortcut. Borrowed from Stephen Turner.
}

