% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outside.R
\name{\%outside\%}
\alias{\%outside\%}
\title{Test if numbers are outside of a specified range}
\usage{
x \%outside\% rng

`\%outside\%`(x, rng, inc_bounds=FALSE)
}
\arguments{
\item{x}{numeric vector to be tested.}

\item{rng}{vector of minimum and maximum values of the range. Elements beyond the first 2 are ignored.}

\item{inc_bounds}{logical, whether to include the bounds in the set that is tested. If set to TRUE, values identical to the bounds are considered matches. Defaults to FALSE. Works only if the function is not used as an infix operator (see usage example below).}
}
\value{
a logical vector with one element for each element in x
}
\description{
Test if a vector of numbers is outside of a specified range, with the bounds optionally included.
}
