% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_outlier.R
\name{is_outlier}
\alias{is_outlier}
\title{Determine outliers}
\usage{
is_outlier(x, na.rm = FALSE)
}
\arguments{
\item{x}{numeric vector containing values to test for outliers}

\item{na.rm}{logical, whether to remove NA values when calculating quantiles and IQR. Passed to \code{quantile()} and \code{IQR()}. Defaults to FALSE.}
}
\value{
logical vector, containing TRUE for outlier elements and FALSE for all other elements.
}
\description{
Determine outliers of a numeric vector, based on the first and third quartile minus/plus 1.5 times the
inter-quartile range.
}
