% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_packages_with_install.R
\name{load_packages_with_install}
\alias{load_packages_with_install}
\title{Load a vector of packages, installing them if they are not already installed}
\usage{
load_packages_with_install(packageNameList, location = "CRAN")
}
\arguments{
\item{packageNameList}{A character vector of package names. For github, the format should be "author/package".}

\item{location}{A character string specifying the location of the packages. Options are "CRAN", "BioConductor", and "github". Partial matches are allowed}
}
\description{
Load a vector of packages. First check if each one is installed, and install
it if not. This function works for CRAN, BioConductor, and github packages.
It can only operate on one type of package at a time. So if you are loading
packages from all three locations, you will need to call this function three
times.
}
