% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logit_transform.R
\name{logit_transform}
\alias{logit_transform}
\title{Transform a vector of proportions using the logit transformation}
\usage{
\code{logit_transform(x, buffer=0.001)}
}
\arguments{
\item{x}{vector of values to transform. Must be between 0 and 1, inclusive.}

\item{buffer}{numeric, the amount to nudge limit values inwards before transformation to avoid returning infinite values. Defaults to 0.001. Set to 0 to return infinite values. Setting to "estimate" will use the smallest distance from 0 or 1 in \code{x} as the buffer.}
}
\value{
a vector of the transformed values
}
\description{
Transform a vector of values between 0 and 1 using the logit transformation.
Includes a buffer, which by default nudges 0s and 1s inward to avoid infinite values.
}
