% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsos.R
\name{lsos}
\alias{lsos}
\title{List the N largest objects in the global environment}
\usage{
lsos(..., n=10)
}
\arguments{
\item{...}{optional arguments, passed to \code{ls.objects}}

\item{n}{integer, the number of objects to list. Defaults to 10.}
}
\description{
Lists the largest objects in the global environment, by memory usage.
}
