% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp.R
\name{lsp}
\alias{lsp}
\title{List package contents}
\usage{
lsp(package, all.names=FALSE, pattern)
}
\arguments{
\item{package}{the package object name (not quoted)}

\item{all.names}{logical, whether to include object names beginning with a '.'. Passed to \code{ls}.}

\item{pattern}{a regular expression, passed to \code{ls}. Only names matching \code{pattern} are returned.}
}
\value{
a vector of object names
}
\description{
List the names of all objects in a package, optionally filtered for pattern matches. Borrowed from @_inundata
}
