% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_copy.R
\name{make_copy}
\alias{make_copy}
\title{Duplicate an object in the global environment}
\usage{
\code{make_copy(var_name, append=".old")}
}
\arguments{
\item{var_name}{character string, the name of the object to be copied.}

\item{append}{character string, the text to be appended to the object's name when making the copy. Defaults to ".old".}
}
\description{
Duplicate an oject in the global environment, so that one version can be modified without losing the previous version. Useful for testing modifications to objects generated by time-consuming code.
}

