% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_transparent.R
\name{make_transparent}
\alias{make_transparent}
\title{Add transparency to a vector of colors}
\usage{
make_transparent(color_names, alpha=100)
}
\arguments{
\item{color_names}{vector of color names to be modified.}

\item{alpha}{numeric, the alpha value, between 0 (completely transparent) and 255 (the original color). Passed to \code{rgb}.}
}
\value{
a character vector of the same length as \code{color_names}
}
\description{
Add a transparency factor to each color in a vector of color names, using alpha on a scale of 0-255. Adapted from http://stackoverflow.com/questions/8047668/transparent-equivalent-of-given-color.
}
