% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeCopies.R
\name{makeCopies}
\alias{makeCopies}
\title{Duplicate objects in the global environment based on names matching a grep pattern.}
\usage{
\code{makeCopies(var_pattern, append=".old", ask=TRUE)}
}
\arguments{
\item{var_pattern}{character string, the regex pattern to match object names. Passed to \code{grep} via \code{ls_grep}.}

\item{append}{character string, the text to be appended to the objects' names when making the copies. Defaults to ".old".}

\item{ask}{logical, whether to prompt the user before copying objects. Defaults to TRUE.}
}
\description{
Duplicate multiple ojects in the global environment based on a grep pattern match. This allows modification of one version of the objects without losing the previous versions. Useful for testing modifications to objects generated by time-consuming code. This is a simple wrapper for makeCopy.
}

