% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/me_labeller.R
\name{me_labeller}
\alias{me_labeller}
\title{Labeller function for plotting WGCNA modules in facet_grid or facet_wrap}
\usage{
\code{facet_grid(.~variable, labeller=me_labeller))}
}
\arguments{
\item{variable}{not used by this function. Included for compatibility with \code{facet_grid} and \code{facet_wrap}.}

\item{value}{character vector containing the levels of the factor used for facetting. Generally passed directly to the function by \code{facet_grid} or \code{facet_wrap}.}
}
\value{
An object of the same dimension as \code{value}, with any "ME" strings removed from each element.
}
\description{
This is a function to pass facet labels to ggplot functions \code{facet_grid} and \code{facet_wrap},
so that the facets have the module color name as the label. Basically, it strips off the "ME" and
returns the label in a format that ggplot understands.
}
