% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/order_points.R
\name{order_points}
\alias{order_points}
\title{Re-order a data frame for plotting points in ggplot}
\usage{
order_points(x, method)
}
\arguments{
\item{x}{data frame containing data to re-order}

\item{method}{character string, specifying how to order the rows. Currently accepted values are "random", which randomizes the order, and "input", which returns the input data frame as-is.}
}
\value{
a data frame with same contents as \code{x}, with rows re-ordered according to \code{method}.
}
\description{
Re-order a data frame to control the plotting order of points in ggplot. This is designed to
control over-plotting of points by sorting the data frame based on input arguments. It is
used in plots by symbol and color, where overplotting may affect readability.
}
