% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_session_info.R
\name{print_session_info}
\alias{print_session_info}
\title{Print session info with better formatting}
\usage{
print_session_info()
}
\description{
Print the session information. This function is a wrapper for \code{sessioninfo::session_info()}, with
the list of packages formatted using \code{kableExtra::kable_styling()} to make it more adaptive to
the output context. In particular, this function works much better with knitr output to html or pdf.
In its current state, there are no parameters, and only default settings. Additional flexibility, 
such as controlling the fields that are output, may be added in the future.
}
