% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo_R2.R
\name{pseudo_R2}
\alias{pseudo_R2}
\title{Calculate a pseudo R^2 value from a fitted model}
\usage{
pseudo_R2(x)
}
\arguments{
\item{x}{result from model fitting, generally the output of lm or lmer.}
}
\value{
numeric, the squared correlation of the fitted values and the observed values of the dependent variable
}
\description{
Calculate the squared correlation of the fitted and observed values for the dependent variable, as a pseudo-R^2. Inputs can be models fit by lmer or lm (other formats to be incorporated later).
}
