% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range01.R
\name{range01}
\alias{range01}
\title{Normalize a numeric vector to 0-1 range}
\usage{
range01(x)
}
\arguments{
\item{x}{numeric vector to be normalized}
}
\description{
This function normalizes a numeric vector to have minimum of 0 and maximum of 1, with values in between linearly
transformed. It ignores NA values in the calculation of min and max.
}
