% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_yn.R
\name{read_yn}
\alias{read_yn}
\title{Prompt user for a yes or no answer
Prompt the user for a yes or no answer, and repeat until an acceptable response is entered}
\usage{
\code{read_yn(display_text)}
}
\arguments{
\item{display_text}{character string, the text to be displayed at the prompt.}
}
\value{
character, either 'y' or 'n'. Uppercase responses are converted to lowercase.
}
\description{
Prompt user for a yes or no answer
Prompt the user for a yes or no answer, and repeat until an acceptable response is entered
}

