% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redundant_cols.model.matrix.R
\name{redundant_cols.model.matrix}
\alias{redundant_cols.model.matrix}
\title{Determine the redundant columns in a model matrix}
\usage{
\code{redundant_cols.model.matrix(model_matrix)}
}
\arguments{
\item{model_matrix}{the model matrix, typically the result of a call to \code{model.matrix} or similar functions.}
}
\value{
a (possibly empty) character vector, with the names of the redundnat columns. Note that these names will typically include the variable name and (if the variable is treated as a factor) the name of the redundant level(s).
}
\description{
This function identifies the redundant columns in a model matrix, if the model matrix is not of full rank.
}
