% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_all_NA_rowcols.R
\name{remove_all_NA_rowcols}
\alias{remove_all_NA_rowcols}
\title{Trim any rows and/or columns containing all NAs from a matrix or data frame.}
\usage{
\code{remove_all_NA_rowcols(x, cols=TRUE, rows=TRUE)}
}
\arguments{
\item{x}{data frame or matrix with rows and/or columns to be removed}

\item{cols, rows}{logical, whether to trim columns and rows, respectively. By default, both columns and rows are trimmed.}

\item{interpret_text}{whether to convert text "NA" to NA values. Default behavior is to set values that are read as "NA" to NA before filtering.}
}
\value{
an object of the same type as \code{x}, potentially with fewer rows and/or columns
}
\description{
Trim any rows and/or columns containing all NAs from a matrix or data frame. Any rows or columns
composed entirely of NAs are removed.This is used primarily after reading in a spreadsheet, when excess
columns and/or rows may be included due to quirks in the file reading process. It can also remove cases
and/or variables with no data.
}
