% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_grep.R
\name{rm_grep}
\alias{rm_grep}
\title{Remove objects matching a regex pattern from global environment}
\usage{
rm_grep(rm_pattern, ask=TRUE, ...)
}
\arguments{
\item{rm_pattern}{character string, the regex pattern to match. Passed to \code{grep}.}

\item{ask}{logical, whether to prompt the user before removing objects. Defaults to TRUE.}

\item{...}{optional arguments passed to \code{grep}.}
}
\description{
Identify and remove all objects in the global environment that match a specified regex pattern. By default, the user is prompted with a list of the objects before they are removed.
}
