% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm_tmp.R
\name{rm_tmp}
\alias{rm_tmp}
\title{Remove temporary objects from global environment}
\usage{
\code{rm_tmp(...)}
}
\arguments{
\item{...}{optional arguments passed to \code{grep} via \code{rm_grep}.}
}
\description{
Clear all temporary variables, after displaying them and (by default) prompting user. By default, the user is prompted with a list of the objects before they are removed. It is assuemd that temporary variables end in ".tmp" or have single-letter names.
}

