% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/se.R
\name{se}
\alias{se}
\title{Calculate the standard error of the mean}
\usage{
\code{se(x)}
}
\arguments{
\item{x}{numeric vector or matrix}

\item{na.rm}{logical. Should missing values be removed?}
}
\value{
a single numeric value
}
\description{
Calculate the standard error of the mean of a vector of numbers, as \code{SD(x) / sqrt(N)}.
}
