% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shuffle_designMat_by_group.R
\name{shuffle_designMat_by_group}
\alias{shuffle_designMat_by_group}
\title{Shuffle a variable in a design matrix within studies or groups}
\usage{
\code{shuffle_designMat_by_group(designMat, shuffle_col, group_cols)}
}
\arguments{
\item{designMat}{a design matrix.}

\item{shuffle_col}{the name or number of the column of \code{designMat} to shuffle.}

\item{group_cols}{the names or numbers of the columns of \code{designMat} corresponding to the groups/studies. Values of the shuffled column are resampled WITHIN group/study}
}
\value{
a design matrix, with dimentions identical to the input \code{designMat}
}
\description{
This function is designed to shuffle annotation data among samples within studies or groups.
It is designed to test the effect of assocations between variables across studies or groups,
by allowing comparison of the trend recovered if all relationships within the data are
retained EXCEPT FOR the values of the variable of interest.
}
