% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sig_digs.R
\name{sig_digs}
\alias{sig_digs}
\title{Calculate the number of significant figures in a number (borrowed from ???)}
\usage{
sig_digs(n)
}
\arguments{
\item{n}{numeric, the number for which to determine significant figures.}
}
\value{
integer, the number of significant figures
}
\description{
Calculate the number of significant figures in an integer or floating point number. NOTE: this function IS NOT vectorized. To run on a vector of numbers, use \code{sapply(x, sigDigs)}. Assumes that consecutive trailing 0s of integers and leading 0s of decimals are not significant. E.g. \code{sigDigs(10) == 1}.
}
