% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simple_cap.R
\name{simple_cap}
\alias{simple_cap}
\title{Capitalize the first letter of each element and/or word in a vector of character strings}
\usage{
\code{simple_cap(x, sep_str=" ")}
}
\arguments{
\item{x}{character vector, with elements to be capitalized.}

\item{sep_str}{character string, the separator between words within vector elements. Vector elements are split on this string, each chunk is capitalized, and they are combined with this string as separator. Defaults to " ".}
}
\value{
a character vector of the same length as \code{x}.
}
\description{
Capitalize the first letter of each word in each element of a character vector. By default, words are assumed to be separated " ".
}
