% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_all_NA_rowcols.R, R/standardize_dimnames.R
\name{standardize_dimnames}
\alias{standardize_dimnames}
\title{Trim any rows and/or columns containing all NAs from a matrix or data frame.}
\usage{
\code{standardize_names(x, cols=TRUE, rows=TRUE, dedup=TRUE)}

\code{standardize_names(x, cols=TRUE, rows=FALSE, dedup=TRUE)}
}
\arguments{
\item{x}{data frame or matrix with rows and/or columns to be removed}

\item{cols, rows}{logical, whether to trim columns and rows, respectively. By default, both columns and rows are trimmed.}

\item{dedup}{logical, whether to deduplicate the vector elements. If TRUE, elements that are identical after standardization are deduplicated by appending "_1", "_2", etc. Defaults to TRUE.}

\item{x}{data frame or matrix with column and/or row names to be standardized}

\item{cols, rows}{logical, whether to standardize the column and row names, respectively. By default, column names are standardized, but row names are not.}
}
\value{
an object of the same type as \code{x}, potentially with fewer rows and/or columns

an object of the same type and dimension as \code{x}
}
\description{
Trim any rows and/or columns containing all NAs from a matrix or data frame. Any rows or columns
composed entirely of NAs are removed.This is used primarily after reading in a spreadsheet, when excess
columns and/or rows may be included due to quirks in the file reading process. It can also remove cases
and/or variables with no data.

Standardize column and/or row names of a matrix or data frame to a standard syntax. This is a wrapper for
\code{standardize_names}, to allow it to be used on data frames and matrices, rather than on their
names. All letters are converted to lowercase, special characters are converted to underscores,
and multiple and trailing (but not leading) special characters are removed. By default, elements are
deduplicated by appending "_1" and so on.
}

