% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_dimnames.R
\name{standardizeDimnames}
\alias{standardizeDimnames}
\title{Standardize the column and/or row names of a matrix or data frame to standard syntax, defaulting to camelCase}
\usage{
standardizeDimnames(
  x, cols = TRUE, rows = FALSE, dedup = TRUE, case = "camelCase")
}
\arguments{
\item{x}{data frame or matrix with column and/or row names to be standardized}

\item{cols, rows}{logical, whether to standardize the column and row names, respectively. By default, column names are standardized, but row names are not.}

\item{dedup}{logical, whether to deduplicate the vector elements. If TRUE, elements that are identical after standardization are deduplicated by appending "_1", "_2", etc. Defaults to TRUE.}

\item{case}{character, specifying whether to convert to camelCase or snake_case. Abbreviations are accepted. Defaults to camelCase.}
}
\value{
an object of the same type and dimension as \code{x}
}
\description{
Standardize column and/or row names of a matrix or data frame to a standard
syntax, defaulting to camelCase. This is a wrapper for 
\code{standardize_names}, to allow it to be used on data frames and
matrices, rather than on their names.
}
