% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values2colors.R
\name{values2colors}
\alias{values2colors}
\title{Convert a vector of numeric values to a vector of colors}
\usage{
\code{values2colors(x,
col.start="blue", col.end="red",
col.pal=NULL, na.col="grey50",
breaks=10)}
}
\arguments{
\item{x}{numeric vector containing values to be converted to colors}

\item{col.start, col.end}{color names indicating the start and end of the color ramp. Passed to \code{colorRampPalette}.}

\item{col.pal}{name of a palette from \code{RColorBrewer}; an alternative way to specify color palette. The specified palette will be passed to \code{colorRampPalette}.}

\item{na.col}{color name to use for NA values in \code{x}. Defaults to "grey50".}

\item{breaks}{number of intervals to break the values of \code{x} into. Defaults to 10.}
}
\value{
a vector of color names, of length equivalent to the length of x.
}
\description{
Convert a vector of numeric values to a vector of color names. The input vector is divided into a number
of intervals, set by \code{breaks}. Values in each interval are assigned a color along a scale generated
by \code{colorRampPalette}.
}
