% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zscore_winsorized.R
\name{zscore_winsorized}
\alias{zscore_winsorized}
\title{Normalize a numeric vector to z-scores with the values winsorized (capped at a maximum absolute value)}
\usage{
zscore_winsorized(x, threshold = 3)
}
\arguments{
\item{x}{numeric vector to be normalized}

\item{threshold}{the maximum absolute value to which the vector will be winsorized}
}
\description{
This function normalizes a numeric vector to z-scores (mean of 0 and standard
deviation of 1), with values beyond a specified threshold winsorized (capped
at a maximum absolute value). It ignores NA values in the calculation of
mean and standard deviation, but returns NAs in the output vector.
}
