% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ls.objects.R
\name{ls.objects}
\alias{ls.objects}
\title{Improved listing of objects (from Dirk Eddelbuettel)}
\usage{
\code{ls.objects(pos = 1, pattern, order.by,
 decreasing=FALSE, head=FALSE, n=5)}
}
\description{
This is an alternative to the base \code{ls} function, with additional functionality. Documentation to be written.
}

