% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range01.R
\name{range01}
\alias{range01}
\title{Normalize a numeric vector to 0-1 range}
\usage{
\code{range01(x)}
}
\arguments{
\item{x}{numeric vector to be normalized}
}
\description{
Normalize a numeric vector to 0-1 range
}

