% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_dates.R
\name{standardize_dates}
\alias{standardize_dates}
\title{Convert dates in various formats to Date format}
\usage{
standardize_dates(x, origin = "1899-12-30", tz = Sys.timezone())
}
\arguments{
\item{x}{vector, the date data to be converted}

\item{origin}{character, the base date for numeric date format, in the form "YYYY-MM-DD". Defaults to "1899-12-30".}

\item{tz}{character, the time zone to be used if converting from POSIX formats. Defaults to PDT, as this package was written for use in a Seattle-based analysis group, and PDT covers more of the year than PST. This affects both the assumed time zone for any dates in character class with "YYYY-MM-DD hh:mm:ss" format, and the time zone used when converting from dateTime format to date alone.}
}
\value{
A vector of the same length as \code{x}
}
\description{
This function is designed to convert dates from a mix of formats to standard "Date" format. It can handle dates in POSIX, numeric, Date, or mixed character format.
}
