% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_dates.R
\name{standardize_dates}
\alias{standardize_dates}
\title{Convert dates in various formats to Date format}
\usage{
standardize_dates(x)
}
\arguments{
\item{x}{vector, the date data to be converted}

\item{origin}{character, the base date for numeric date format, in the form "YYYY-MM-DD". Defaults to "1899-12-30".}
}
\value{
A vector of the same length as \code{x}
}
\description{
This function is designed to convert dates from a mix of formats to standard "Date" format.
frames, data frame like objects (e.g. tibbles), or matrices. It returns an object of the same class.
The values that will be coerced to NAs can be provided as a character vector.
}
