% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize_names.R
\name{standardize_names}
\alias{standardize_names}
\title{Standardize a character vector to standard syntax (deduplication copied from limma::makeUnique)}
\usage{
\code{standardize_names(x, dedup=TRUE)}
}
\arguments{
\item{x}{character vector with elements to be standardized}

\item{dedup}{logical, whether to deduplicate the vector elements. If TRUE, elements that are identical after standardization are deduplicated by appending "_1", "_2", etc. Defaults to TRUE.}
}
\value{
a character vector of the same length as \code{x}
}
\description{
Standardize all elements of a character vector to the same syntax. All letters are converted to lowercase, special characters are converted to underscores, and multiple and trailing (but not leading) special characters are removed. By default, elements are deduplicated by appending "_1" and so on.
}

